+++ src/main/java/com/floney/floney/alarm/repository/AlarmCustomRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.floney.floney.infra.alarm.repository;
+package com.floney.floney.alarm.repository;
 
 import com.floney.floney.book.entity.Book;
 import com.floney.floney.book.entity.BookUser;
